Remove -L 1 xargs switch
With recent versions of xargs the program complains about the following:
> xargs: warning: options --max-lines and --replace/-I/-i are mutually
>        exclusive, ignoring previous --max-lines value

This message makes no sense, because as per the documentation --replace
*implies* -L 1.
With this change we remove this switch to silence the program.
Furthermore, the long option --replace has been deprecated, with the
only remaining replacement being a short option: -I. With this change we
switch over to using it instead. Good lord.